fix: create new component from chat to chatheader
