+++ src/pages/Chat/Chat.tsx
@@ -8,26 +8,11 @@ import {
 import ChatbotThinking from '@/components/ChatbotThinking/ChatbotThinking';
 import AuthService from '@/services/authService';
 import { chatService } from '@/services/chatService';
-import {
-  AddCircleOutline,
-  Language,
-  LockOutline,
-  Logout,
-  Menu as MenuIcon,
-} from '@mui/icons-material';
-import {
-  Button,
-  CircularProgress,
-  ListItemIcon,
-  ListItemText,
-  Menu,
-  MenuItem,
-} from '@mui/material';
-import FindInPageIcon from '@mui/icons-material/FindInPage';
+import { CircularProgress } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
-import ChatbotForm from '../../components/Form/ChatbotForm';
-import ChatbotMessage, { IChat } from '../../components/Message/ChatbotMessage';
-import ChatbotIcon from '../../components/icons/ChatbotIcon';
+import ChatbotForm from '@/components/Form/ChatbotForm';
+import ChatHeader from '@/components/Header/ChatHeader';
+import ChatbotMessage, { IChat } from '@/components/Message/ChatbotMessage';
 import Styles from './_Chat.module.scss';
 
 const Chat = () => {
@@ -37,14 +22,11 @@ const Chat = () => {
 
   const [folioNumber, setFolioNumber] = useState('');
   const [loading, setLoading] = useState(true);
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const [isThinking, setIsThinking] = useState(false);
-  const { language, setLanguage, chatHistory, setChatHistory } =
-    useAppContext();
+  const { chatHistory, setChatHistory } = useAppContext();
   const chatBodyRef = useRef<HTMLDivElement>(null);
   const hasInitializedRef = useRef(false);
   const { setAlert } = useAlert();
-  const open = Boolean(anchorEl);
 
   useEffect(() => {
     if (hasInitializedRef.current) return;
@@ -102,33 +84,7 @@ const Chat = () => {
     });
   }, [chatHistory]);
 
-  const toggleLanguage = () => {
-    const newLang = language === 'en' ? 'es' : 'en';
-    setLanguage(newLang);
-  };
-
   const lastBotIndex = chatHistory?.map((m) => m.role).lastIndexOf('bot');
-
-  const handleLogout = async () => {
-    try {
-      await AuthService.logout();
-      localStorage.removeItem('token');
-      setUser(null);
-    } catch (error) {
-      localStorage.removeItem('token');
-      setUser(null);
-    } finally {
-      navigate('/');
-    }
-  };
-
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
 
   const handleNewConversation = () => {
     localStorage.removeItem('conversationId');
@@ -139,88 +95,10 @@ const Chat = () => {
     <div className={Styles.container}>
       <div className={Styles.chat}>
         {/* Chatbot Header */}
-        <header className={Styles.chat_header}>
-          <section className={Styles.chat_header_top}>
-            <div className={Styles.header_info}>
-              <ChatbotIcon />
-              <h2 className={Styles.logo_text}>{t.header}</h2>
-            </div>
-
-            <div className={Styles.new_chat_container}>
-              <div className={Styles.new_chat}>
-                <Button onClick={handleNewConversation}>
-                  <p className={Styles.new_chat_text}>
-                    {t.menu.new_conversation}
-                  </p>
-                </Button>
-              </div>
-
-              <div className={Styles.buttons_header}>
-                <Button
-                  id="basic-button"
-                  aria-controls={open ? 'basic-menu' : undefined}
-                  aria-haspopup="true"
-                  aria-expanded={open ? 'true' : undefined}
-                  onClick={handleClick}
-                >
-                  <MenuIcon />
-                </Button>
-                <Menu
-                  id="basic-menu"
-                  anchorEl={anchorEl}
-                  open={open}
-                  onClose={handleClose}
-                  slotProps={{
-                    list: {
-                      'aria-labelledby': 'basic-button',
-                    },
-                  }}
-                >
-                  <MenuItem
-                    onClick={() => {
-                      handleClose();
-                      navigate('/change-password');
-                    }}
-                  >
-                    <ListItemIcon>
-                      <LockOutline fontSize="small" />
-                    </ListItemIcon>
-                    <ListItemText>{t.menu.change_password}</ListItemText>
-                  </MenuItem>
-                  <MenuItem onClick={toggleLanguage}>
-                    <ListItemIcon>
-                      <Language fontSize="small" />
-                    </ListItemIcon>
-                    <ListItemText>{t.menu.lang}</ListItemText>
-                  </MenuItem>
-
-                  <MenuItem
-                    onClick={() => {
-                      handleClose();
-                      navigate('/recovery');
-                    }}
-                  >
-                    <ListItemIcon>
-                      <FindInPageIcon fontSize="small" />
-                    </ListItemIcon>
-                    <ListItemText>{t.menu.recovery}</ListItemText>
-                  </MenuItem>
-
-                  <MenuItem onClick={handleLogout}>
-                    <ListItemIcon>
-                      <Logout fontSize="small" />
-                    </ListItemIcon>
-                    <ListItemText>{t.menu.logout}</ListItemText>
-                  </MenuItem>
-                </Menu>
-              </div>
-            </div>
-          </section>
-
-          <div className={Styles.folio_number}>
-            {folioNumber && <p>{`${t.folio}: ${folioNumber}`}</p>}
-          </div>
-        </header>
+        <ChatHeader
+          folioNumber={folioNumber}
+          onNewConversation={handleNewConversation}
+        />
 
         {/* Chatbot Body */}
         <div className={Styles.chat_body} ref={chatBodyRef}>
